+++ src/tests/ftest/pool/evict_test.py
@@ -260,7 +260,7 @@ class EvictTests(TestWithServers):
         """
         Test evicting a pool using an invalid server group name.
 
-        :avocado: tags=all,pool,pr,full_regression,small,poolevict,
+        :avocado: tags=all,pool,pr,full_regression,small,poolevict
         :avocado: tags=poolevict_bad_server_name
         """
         test_param = self.params.get("server_name", '/run/badparams/*')
@@ -270,7 +270,7 @@ class EvictTests(TestWithServers):
         """
         Test evicting a pool using an invalid uuid.
 
-        :avocado: tags=all,pool,pr,full_regression,small,poolevict,
+        :avocado: tags=all,pool,pr,full_regression,small,poolevict
         :avocado: tags=poolevict_bad_uuid
         """
         test_param = self.params.get("uuid", '/run/badparams/*')
